pyproject.toml
[project]
name = "yuki-chan-bot"
version = "8.0.0"
description = "Yuki Chan Bot - A cute QQ bot powered by LLM and RAG"
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "aiohttp>=3.9.0",
    "websockets>=14.0",
    "openai>=1.50.0",
    "tenacity>=8.2.0",
    "chromadb>=0.5.0",
    "sentence-transformers>=2.2.0",
    "numpy>=1.24.0",
    "jieba>=0.42.1",
    "opencv-python>=4.8.0",
    "pyyaml>=6.0",
    "ollama>=0.4.0",
]

[project.optional-dependencies]
dev = []

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["."]

[tool.uv]
package = true
managed = true
